refactor: move files to trading dir
