+++ package.json
@@ -14,7 +14,7 @@
   },
   "types": "dist/index.d.ts",
   "scripts": {
-    "prebuild": "rm -rf dist && yarn run codegen",
+    "prebuild": "rm -rf dist",
     "build": "microbundle -f modern,esm,cjs",
     "start": "microbundle -f modern,esm,cjs watch",
     "postbuild": "cp package.json dist && cp README.md dist",
@@ -31,6 +31,7 @@
     "typechain:codegen": "typechain --target ethers-v5 --out-dir ./src/composable/generated './abi/*.json'"
   },
   "dependencies": {
+    "@cowprotocol/app-data": "^2.1.0",
     "@cowprotocol/contracts": "^1.6.0",
     "@ethersproject/abstract-signer": "^5.7.0",
     "@openzeppelin/merkle-tree": "^1.0.5",
