+++ src/index.ts
@@ -3,3 +3,4 @@ export * from './order-book'
 export * from './subgraph'
 export * from './order-signing'
 export * from './composable'
+export * from './trading'
