+++ src/trading/appDataUtils.ts
